+++ client/Combatant/CombatantViewModel.ts
@@ -1,3 +1,4 @@
+import { toModifierString } from "../../common/Toolbox";
 import { CombatantCommander } from "../Commands/CombatantCommander";
 import { ConcentrationPrompt } from "../Commands/Prompts/ConcentrationPrompt";
 import { DefaultPrompt, Prompt } from "../Commands/Prompts/Prompt";
@@ -5,7 +6,6 @@ import { TagPrompt } from "../Commands/Prompts/TagPrompt";
 import { Encounter } from "../Encounter/Encounter";
 import { CurrentSettings } from "../Settings/Settings";
 import { Metrics } from "../Utility/Metrics";
-import { toModifierString } from "../Utility/Toolbox";
 import { Combatant } from "./Combatant";
 import { Tag } from "./Tag";
