+++ client/Commands/Prompts/InitiativePrompt.ts
@@ -1,6 +1,6 @@
+import { toModifierString } from "../../../common/Toolbox";
 import { Combatant } from "../../Combatant/Combatant";
 import { TutorialSpy } from "../../Tutorial/TutorialViewModel";
-import { toModifierString } from "../../Utility/Toolbox";
 import { Prompt } from "./Prompt";
 
 export class InitiativePrompt implements Prompt {
