+++ client/Utility/CustomBindingHandlers.ts
@@ -1,4 +1,5 @@
 import { render as renderReact } from "react-dom";
+import { toModifierString } from "../../common/Toolbox";
 import { CombatantCommander } from "../Commands/CombatantCommander";
 import { EncounterCommander } from "../Commands/EncounterCommander";
 import { SpellLibrary } from "../Library/SpellLibrary";
@@ -6,7 +7,6 @@ import { Dice, IRules } from "../Rules/Rules";
 import { TrackerViewModel } from "../TrackerViewModel";
 import { ComponentLoader } from "./Components";
 import { TextAssets } from "./TextAssets";
-import { toModifierString } from "./Toolbox";
 
 declare var markdownit: any;
 declare var Awesomplete: any;
